Remove code that subtracts old key/value size from memtable when replacing.
This is a tradeoff between the size of the journal file and the size of the SortedBlockTable file when repeatedly setting the same key over and over in the database. Having a bounded size on the Journal file is more important than making sure that SortedBlockTable files are a consistently large size. In any case, having a wider distribution of keys is more efficient than resetting the same key over and over.