Put an upper boundary on the worst-case merge operation.
SBT's need to roll over when they reach a key that represents a 10 page span in the level above them, even if they haven't reached the size limit on the current merge page. This puts an upper bound of 12 on the number of pages included in any single merge operation. If we don't do this, then high key density in mid-level pages can trigger huge merges with an unbounded number of upper level pages. I've observed cases with 100+ pages in a single merge. This wreaks havoc with disk I/O and kills performance on the local machine.